requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile
#
certifi==2024.2.2
    # via requests
cffi==1.16.0
    # via
    #   sounddevice
    #   vosk
charset-normalizer==3.3.2
    # via requests
idna==3.7
    # via requests
numpy==1.26.4
    # via -r requirements.in
pycparser==2.22
    # via cffi
pyyaml==6.0.1
    # via -r requirements.in
requests==2.31.0
    # via vosk
sounddevice==0.4.6
    # via -r requirements.in
srt==3.5.3
    # via vosk
tqdm==4.66.2
    # via vosk
urllib3==2.2.1
    # via requests
vosk==0.3.44
    # via -r requirements.in
websockets==12.0
    # via vosk
